change Datatype FLOAT to DECIMAL

+++ src/models/Size.ts
@@ -11,12 +11,12 @@ const Size: ModelStatic<SizeModel> = sequelize.define<SizeModel>('sizes', {
     field: 'size_id'
   },
   extra: {
-    type: DataTypes.FLOAT(10, 2),
+    type: DataTypes.DECIMAL(10, 2),
     defaultValue: 0,
     allowNull: false
   },
   size: {
-    type: DataTypes.FLOAT(2, 1),
+    type: DataTypes.DECIMAL(2, 1),
     allowNull: false
   },
   stock: {
